packaging and some errors handled

+++ src/main/java/az/edu/turing/happyfamilyv1/dao/FamilyDao.java
@@ -21,7 +21,7 @@ public interface FamilyDao {
 
     List<Family> getFamiliesLessThan(int peopleCount);
 
-    int countFamiliesWithMemberNumber(int peopleCount);
+    long countFamiliesWithMemberNumber(long peopleCount);
 
     Family adoptChild(Family family, Human child);
